still working on hal
