--- src/Kernal/BUILD.h
@@ -1,1 +1,0 @@
-#define BUILD 1796
