--- src/Kernal/Kernal.h
@@ -1,10 +1,0 @@
-#ifndef KERNAL_H_
-#define KERNAL_H_
-
-
-#include "../lib/IO/Keyboard.h"
-
-
-
-
-#endif // !KERNAL_H_
