+++ src/Kernel/CppKern.h
@@ -1,0 +1,14 @@
+#ifndef KERN_SRT_H
+#define KERN_SRT_H
+
+#include "../lib/core/core.h"
+#include "../lib/Term/Term.h"
+
+EXTNC_ int KernStart();
+
+EXTNC_ int tellout(char * str);
+
+
+#endif // !KERN_SRT_H
+
+
