+++ src/Kernel/Kernel.h
@@ -1,0 +1,10 @@
+#ifndef KERNAL_H_
+#define KERNAL_H_
+
+
+#include "../lib/IO/Keyboard.h"
+
+
+
+
+#endif // !KERNAL_H_
