+++ src/lib/hal/hal.cpp
@@ -1,0 +1,9 @@
+#include "hal.h"
+
+void SYS::reset() {
+    uint8_t good = 0x02;
+    while (good & 0x02)
+        good = port_byte_in(0x64);
+    port_byte_out(0x64, 0xFE);
+    HALT;
+}
